about to try out redux-modal
