+++ lambda-notes/package.json
@@ -14,6 +14,7 @@
     "reactstrap": "^5.0.0",
     "redux": "^4.0.0",
     "redux-logger": "^3.0.6",
+    "redux-modal": "^1.7.3",
     "redux-thunk": "^2.2.0"
   },
   "scripts": {
