+++ lambda-notes/src/components/Modal/DeleteModal.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button, Modal, ModalHeader, ModalBody, ModalFooter } from 'reactstrap';
+import { Button, Modal, ModalBody, ModalFooter } from 'reactstrap';
 
 class DeleteModal extends React.Component {
   constructor(props) {
@@ -7,11 +7,9 @@ class DeleteModal extends React.Component {
     this.state = {
       modal: false
     };
-
-    this.toggle = this.toggle.bind(this);
   }
 
-  toggle() {
+  toggle = () => {
     this.setState({
       modal: !this.state.modal
     });
