+++ lambda-notes/src/reducers/ModalReducer.js
@@ -1,0 +1,20 @@
+import { SHOW_MODAL, HIDE_MODAL } from '../actions';
+
+const initialState = {
+    modalType: null,
+    modalProps: {}
+}
+
+export default (state = initialState, action) => {
+    switch (action.type) {
+        case SHOW_MODAL:
+            return {
+                modalType: action.modalType,
+                modalProps: action.modalProps
+            }
+        case HIDE_MODAL:
+            return initialState
+        default:
+            return state
+    }
+}
